<3ThomasBS_ILoveYOU!! [ Windows Desktop : day 90 ] Building application at design solution <3ThomasBS_ILoveYOU!!
<3tbs, Mummi and bilbo!!

Thomas Borregaard Sørensen \infinity,-0.16091989,\infinity ONE-MAN ABSOLUTE <3!! I love you, by ???-0.02041977-???write my history please make me please create me for you for me for you for me Camilo Sasuke Thomas Borregaard Sørensen!!

Thomas 3 private commits on mid Dec2020!!

Thomas Online YouTube VODs contribution!!

Mummi orange-rice-flour cake on 20-Dec!!

Mummi (tinytaura) watching and chatting contribution!!

bilbo sleeping and needing/requesting/crying for help care (for the right person (me), the cats wanna fight with him) contribution!!

sodapoppin and friends contribution!!

iAssyrian chatting contribution!!

boflux (Spoofh, Benjamin Kuhl) chatting contribution!!

jusg_fpga (fpga_guru, vue_equalizer, just_fpga, Oliver Pohl) chatting contribution!!

cmgriffing streaming contribution!!

TimBeaudet (Friends: FletcherLabs, tsjost and Jabokoe) streaming contribution!!

Stumpen_nicklas_dk, sodapoppin and EduardoRFS streaming contribution!!

Roxkstar74 sleeping streaming contribution!!

kissloryshy chatting contribution!!

blackjekko from Padova Italia through twitch C++/ca2 interest contribution!!

j_blow streaming contribution!!

boflux (Ben, Spoofh, from Germany) chatting contribution!!

parrot_rl chatting contribution (from New Jersey)!!

JPCdk streaming contribution!!

whyyyyyyysoserious streaming chess contribution!!

fpga_guru (vue_equalizer, Oliver from Deutsch)  C++/ca2 interest contribution!!

SovereignDev with Unreal streaming contribution!!

Ash_F0x and TimBeaudet streaming contribution!!

Myrkee (Valheim) streaming contribution!!

Knut streaming contribution!!

shegotmad (from London) chatting contriubtion!!

dandymcgee (from US) chatting and streaming contribution!!

xmetrix and EinfachUwe42 streaming contribution!!

JessicaMak and marcobrunodev streaming contribution!!

alfredotigolo, mandrakenk and Okbatgames chatting contribution!!

jitspoe, Endesga and Fearitself streaming contribution!!

jmcmorris (Jason Morris, SiegeGames) streaming contribution!!

tomrandall streaming Ludum contribution!!

vue_equalizer (fpga_guru) chatting contribution!!

Thiagovgamg chatting contribution!!

Naysayer88 and friends contribution!!

wtfblub (@Twitch) actively helping ca2 with good OpenGL hints!!

lelandkwong streaming contribution!!

Goldbargames streaming contribution!!

Bytakos (bytakos) streaming contribution!!

Endesga streaming contribution!!

jitspoe and strager streaming contribution!!

Ash_F0x and JessicaMak streaming contribution!!

WTSRetro/SpiffyDane and Myrkee streaming contribution!!

Ninja and friends streaming contribution!!

erald_guri chatting contribution!!

lastmiles streaming farwest contribution!!

rw_grim streaming contribution!!

AdamCYounis streaming contribution!!

Dunno (P4ndaExpress) chatting and streaming contribution!!

Zorchenhimer streaming contribution!!

lasteveq4 C++ interest chat contriubtion!!

cecilphillip and clarkio @"Microsoft Developer" streaming contribution!!

oijtx streaming contribution!!

diegobrando_linux (Bl4ck_gookoo) chatting contribution!!

jhovgaard streaming contribution!!

Klay4_ chatting contribution!!

HonestDanGames streaming contribution!!

NorthSeaHero streaming contribution!!

Trainwreckstv and friends streaming contribution!!

togglebit, GexYT and GoPirateSoftware streaming contribution!!

taiyoinoue, RetroMMO, OfficialAndyPyro and david_joffe streaming contribution!!

Tjienta streaming contribution!!

Primeagen streaming contribution!!

Jaxstyle and friends streaming contribution!!

EduardRFS streaming contribution!!

Melchizedek6809 and btcfly streaming contribution!!

Llama0x0 and sov_l chatting contribution!!

TaleLearnCode streaming contribution!!

Carol phone call contribution and visit contribution!!

hvalen_hvalborg112 streaming contribution!!

harmannieves chatting contribution!! (After long time...)

darkfolt8 (French from France) chatting contribution!!

klintcsgo (CS GO: Counter-Strike Global Offensive) streaming contribution!!

KASPERPURE (Super Mario 64) streaming contribution!!

SomewhatAccurate C++ streaming contribution!!

Listening to Bryan Adams, Westlife, Shayne Ward, MLTR, Backstreet Boys, Boyzone - Best Love Songs Ever by Relax Song at YouTube!!

-- hi5 contribution...!!

at Windows 10 Pro at France remotely from bilbo machine running Windows 10 Pro!! dedicated server by OVH.com at France, Gravelines
Intel Core i7-4790K - 4c/8t - 4 GHz/4.4 GHz RAM32 GB 1600 MHz 2×960 GB SSD SATA

+++ draw2d_opengl/line1.h
@@ -76,14 +76,14 @@ void line( double x1, double y1, double x2, double y2, //coordinates of the line
 	/*   */if ( w>=0.0f && w<1.0f) {
 		t=0.05f; R=0.48f+0.32f*f;
 		if ( !alphablend) {
-			Cr+= 0.88f*(1-f);
-			Cg+= 0.88f*(1-f);
-			Cb+= 0.88f*(1-f);
+			Cr+= 0.88f*(1.0f - (float)f);
+			Cg+= 0.88f*(1.0f - (float)f);
+			Cb+= 0.88f*(1.0f - (float)f);
 			if ( Cr>1.0f) Cr=1.0f;
 			if ( Cg>1.0f) Cg=1.0f;
 			if ( Cb>1.0f) Cb=1.0f;
 		} else {
-			A*=f;
+			A*=(float)f;
 		}
 	} else if ( w>=1.0 && w<2.0) {
 		t=0.05f+f*0.33f; R=0.768f+0.312f*f;
@@ -164,19 +164,19 @@ void line( double x1, double y1, double x2, double y2, //coordinates of the line
 	//draw the line by triangle strip m:1
 glBegin(GL_TRIANGLE_STRIP);
 	if ( !alphablend) {glColor3f( Br,Bg,Bb);} else {glColor4f( Cr,Cg,Cb, 0);}
-	glVertex2f( x1-tx-Rx-cx, y1-ty-Ry-cy); //fading edge
-	glVertex2f( x2-tx-Rx+cx, y2-ty-Ry+cy);
+	glVertex2f((float)(x1-tx-Rx-cx),(float)(y1-ty-Ry-cy)); //fading edge
+	glVertex2f((float)(x2-tx-Rx+cx),(float)(y2-ty-Ry+cy));
 	
 	if ( !alphablend) {glColor3f( Cr,Cg,Cb);} else {glColor4f( Cr,Cg,Cb, A);}
-	glVertex2f( x1-tx-cx,y1-ty-cy); //aura
-	glVertex2f( x2-tx+cx,y2-ty+cy);
-	glVertex2f( x1+tx-cx,y1+ty-cy);
-	glVertex2f( x2+tx+cx,y2+ty+cy);
+	glVertex2f((float)(x1-tx-cx), (float)(y1-ty-cy)); //aura
+	glVertex2f((float)(x2-tx+cx), (float)(y2-ty+cy));
+	glVertex2f((float)(x1+tx-cx), (float)(y1+ty-cy));
+	glVertex2f((float)(x2+tx+cx), (float)(y2+ty+cy));
 	
 	if ( !alphablend) {glColor3f( Br,Bg,Bb);} else {glColor4f( Cr,Cg,Cb, 0);}
-	glVertex2f( x1+tx+Rx-cx, y1+ty+Ry-cy); //fading edge
-	glVertex2f( x2+tx+Rx+cx, y2+ty+Ry+cy);
-glEnd();
+	glVertex2f((float)(x1+tx+Rx-cx), (float)(y1+ty+Ry-cy)); //fading edge
+	glVertex2f((float)(x2+tx+Rx+cx), (float)(y2+ty+Ry+cy));
+   glEnd();
 
 	//cap
 	if ( w < 3) {
@@ -185,29 +185,29 @@ glEnd();
 		//draw cap
 	glBegin(GL_TRIANGLE_STRIP);
 		if ( !alphablend) {glColor3f( Br,Bg,Bb);} else {glColor4f( Cr,Cg,Cb, 0);}
-		glVertex2f( x1-tx-Rx-cx, y1-ty-Ry-cy);
-		glVertex2f( x1-tx-Rx, y1-ty-Ry);
+		glVertex2f((float)(x1-tx-Rx-cx), (float)(y1-ty-Ry-cy));
+		glVertex2f((float)(x1-tx-Rx), (float)(y1-ty-Ry));
 		if ( !alphablend) {glColor3f( Cr,Cg,Cb);} else {glColor4f( Cr,Cg,Cb, A);}
-		glVertex2f( x1-tx-cx, y1-ty-cy);
+		glVertex2f((float)(x1-tx-cx), (float)(y1-ty-cy));
 		if ( !alphablend) {glColor3f( Br,Bg,Bb);} else {glColor4f( Cr,Cg,Cb, 0);}
-		glVertex2f( x1+tx+Rx, y1+ty+Ry);
+		glVertex2f((float)(x1+tx+Rx), (float)(y1+ty+Ry));
 		if ( !alphablend) {glColor3f( Cr,Cg,Cb);} else {glColor4f( Cr,Cg,Cb, A);}
-		glVertex2f( x1+tx-cx, y1+ty-cy);
+		glVertex2f((float)(x1+tx-cx), (float)(y1+ty-cy));
 		if ( !alphablend) {glColor3f( Br,Bg,Bb);} else {glColor4f( Cr,Cg,Cb, 0);}
-		glVertex2f( x1+tx+Rx-cx, y1+ty+Ry-cy);
+		glVertex2f((float)(x1+tx+Rx-cx), (float)(y1+ty+Ry-cy));
 	glEnd();
 	glBegin(GL_TRIANGLE_STRIP);
 		if ( !alphablend) {glColor3f( Br,Bg,Bb);} else {glColor4f( Cr,Cg,Cb, 0);}
-		glVertex2f( x2-tx-Rx+cx, y2-ty-Ry+cy);
-		glVertex2f( x2-tx-Rx, y2-ty-Ry);
+		glVertex2f((float)(x2-tx-Rx+cx), (float)(y2-ty-Ry+cy));
+		glVertex2f((float)(x2-tx-Rx), (float)(y2-ty-Ry));
 		if ( !alphablend) {glColor3f( Cr,Cg,Cb);} else {glColor4f( Cr,Cg,Cb, A);}
-		glVertex2f( x2-tx+cx, y2-ty+cy);
+		glVertex2f((float)(x2-tx+cx), (float)(y2-ty+cy));
 		if ( !alphablend) {glColor3f( Br,Bg,Bb);} else {glColor4f( Cr,Cg,Cb, 0);}
-		glVertex2f( x2+tx+Rx, y2+ty+Ry);
+		glVertex2f((float)(x2+tx+Rx), (float)(y2+ty+Ry));
 		if ( !alphablend) {glColor3f( Cr,Cg,Cb);} else {glColor4f( Cr,Cg,Cb, A);}
-		glVertex2f( x2+tx+cx, y2+ty+cy);
+		glVertex2f((float)(x2+tx+cx), (float)(y2+ty+cy));
 		if ( !alphablend) {glColor3f( Br,Bg,Bb);} else {glColor4f( Cr,Cg,Cb, 0);}
-		glVertex2f( x2+tx+Rx+cx, y2+ty+Ry+cy);
+		glVertex2f((float)(x2+tx+Rx+cx), (float)(y2 + ty + Ry + cy));
 	glEnd();
 	}
 }
@@ -258,19 +258,19 @@ void hair_line( double x1, double y1, double x2, double y2, bool alphablend=0)
 	
 	//draw the line by triangle strip
 glBegin(GL_TRIANGLE_STRIP);
-	if ( !alphablend) {glColor3f( 1,1,1);} else {glColor4f( C,C,C, 0);}
-	glVertex2f( x1-tx-Rx, y1-ty-Ry); //fading edge
-	glVertex2f( x2-tx-Rx, y2-ty-Ry);
+	if ( !alphablend) {glColor3f( 1.0f,1.0f,1.0f);} else {glColor4f((float) C, (float)C, (float)C, 0.0f);}
+	glVertex2f((float)(x1-tx-Rx), (float)(y1-ty-Ry)); //fading edge
+	glVertex2f((float)(x2-tx-Rx), (float)(y2-ty-Ry));
 	
-	glColor3f( C,C,C);	
-	glVertex2f( x1-tx,y1-ty); //aura
-	glVertex2f( x2-tx,y2-ty);
-	glVertex2f( x1+tx,y1+ty);
-	glVertex2f( x2+tx,y2+ty);
+	glColor3f((float)(C), (float)(C), (float)(C));
+	glVertex2f((float)(x1-tx), (float)(y1-ty)); //aura
+	glVertex2f((float)(x2-tx), (float)(y2-ty));
+	glVertex2f((float)(x1+tx), (float)(y1+ty));
+	glVertex2f((float)(x2+tx), (float)(y2+ty));
 	
-	if ( !alphablend) {glColor3f( 1,1,1);} else {glColor4f( C,C,C, 0);}
-	glVertex2f( x1+tx+Rx, y1+ty+Ry); //fading edge
-	glVertex2f( x2+tx+Rx, y2+ty+Ry);
+	if ( !alphablend) {glColor3f( 1.f,1.f,1.f);} else {glColor4f((float) C, (float)C, (float)C, 0.f);}
+	glVertex2f((float)(x1+tx+Rx), (float)(y1+ty+Ry)); //fading edge
+	glVertex2f((float)(x2+tx+Rx), (float)(y2+ty+Ry));
 glEnd();
 }
 
@@ -280,11 +280,11 @@ void line_raw( double x1, double y1, double x2, double y2,
 	double Cr, double Cg, double Cb,
 	double,double,double, bool)
 {
-	glLineWidth(w);
+	glLineWidth((float)(w));
 glBegin(GL_LINES);
-	glColor3f( Cr,Cg,Cb);
-	glVertex2f( x1,y1);
-	glVertex2f( x2,y2);
+	glColor3f((float)(Cr), (float)(Cg), (float)(Cb));
+	glVertex2f((float)(x1), (float)(y1));
+	glVertex2f((float)(x2), (float)(y2));
 glEnd();
 }
